+++ utils/restore_thethe_db.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 # https://jeromejaglale.com/doc/programming/mongodb_docker_mongodump_mongorestore
 
+if [ "$1" == "" ]; then
+    echo "Usage: $0 <name_of_mongodb_container>"
+    exit 1
+fi
+
 gzip -kd db.dump.gz
-docker exec -i thethe_mongo_1 sh -c 'mongorestore --username root  --password root --authenticationDatabase admin --archive' < db.dump
+docker exec -i $1 sh -c 'mongorestore --drop --username root  --password root --authenticationDatabase admin --archive' < db.dump
 rm db.dump
